reorganized testing...
Had to completely reorganize tests to make it sane before adding in a test for the modified `add_pca3` method. The structure of the testing module is much better since test functions are broken into submodules matching the structure of the actual lipydomics modules.
