+++ lipydomics/doc/lipydomics.test.__main__.html
@@ -14,6 +14,26 @@ Dylan&nbsp;H.&nbsp;Ross<br>
 2019/02/02<br>
 &nbsp;<br>
 description:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;Calls&nbsp;the&nbsp;run_all_tests&nbsp;method&nbsp;to&nbsp;run&nbsp;all&nbsp;of&nbsp;the&nbsp;tests&nbsp;defined&nbsp;in&nbsp;tests.py</tt></p>
-
+&nbsp;&nbsp;&nbsp;&nbsp;Calls&nbsp;the&nbsp;run_all_tests&nbsp;method&nbsp;to&nbsp;run&nbsp;all&nbsp;of&nbsp;the&nbsp;tests&nbsp;defined&nbsp;in&nbsp;the&nbsp;various&nbsp;testing&nbsp;modules</tt></p>
+<p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#aa55cc">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="os.html">os</a><br>
+</td><td width="25%" valign=top><a href="traceback.html">traceback</a><br>
+</td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#eeaa77">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Functions</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#eeaa77"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><dl><dt><a name="-run_all_tests"><strong>run_all_tests</strong></a>()</dt><dd><tt>run_all_tests<br>
+&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;runs&nbsp;all&nbsp;tests&nbsp;sequentially,&nbsp;if&nbsp;there&nbsp;are&nbsp;any&nbsp;failures&nbsp;the&nbsp;test&nbsp;function&nbsp;docstring&nbsp;is&nbsp;printed.&nbsp;A&nbsp;traceback&nbsp;is&nbsp;<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;printed&nbsp;as&nbsp;well&nbsp;if&nbsp;an&nbsp;exception&nbsp;is&nbsp;thrown.</tt></dd></dl>
+</td></tr></table>
 </body></html>
