+++ lipydomics/doc/lipydomics.test.data.html
@@ -1,20 +1,20 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module lipydomics.test.tests</title>
+<html><head><title>Python: module lipydomics.test.data</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 </head><body bgcolor="#f0f0f8">
 
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
 <tr bgcolor="#7799ee">
 <td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong><a href="lipydomics.html"><font color="#ffffff">lipydomics</font></a>.<a href="lipydomics.test.html"><font color="#ffffff">test</font></a>.tests</strong></big></big></font></td
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong><a href="lipydomics.html"><font color="#ffffff">lipydomics</font></a>.<a href="lipydomics.test.html"><font color="#ffffff">test</font></a>.data</strong></big></big></font></td
 ><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/tests.py">/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/tests.py</a></font></td></tr></table>
-    <p><tt>lipydomics/test/tests.py<br>
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/data.py">/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/data.py</a></font></td></tr></table>
+    <p><tt>lipydomics/test/data.py<br>
 Dylan&nbsp;H.&nbsp;Ross<br>
-2019/02/02<br>
+2019/09/25<br>
 &nbsp;<br>
 description:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;TODO</tt></p>
+&nbsp;&nbsp;&nbsp;&nbsp;Define&nbsp;tests&nbsp;associated&nbsp;with&nbsp;the&nbsp;lipydomics/data.py&nbsp;module</tt></p>
 <p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#aa55cc">
@@ -24,8 +24,7 @@ description:<br>
 <tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="numpy.html">numpy</a><br>
 </td><td width="25%" valign=top><a href="os.html">os</a><br>
-</td><td width="25%" valign=top><a href="traceback.html">traceback</a><br>
-</td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+</td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#eeaa77">
 <td colspan=3 valign=bottom>&nbsp;<br>
@@ -53,22 +52,6 @@ description:<br>
 &nbsp;&nbsp;&nbsp;&nbsp;description:<br>
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;TODO<br>
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Test&nbsp;fails&nbsp;if<br>
-&nbsp;&nbsp;&nbsp;&nbsp;returns:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(bool)&nbsp;--&nbsp;test&nbsp;pass&nbsp;(True)&nbsp;or&nbsp;fail&nbsp;(False)</tt></dd></dl>
- <dl><dt><a name="-run_all_tests"><strong>run_all_tests</strong></a>()</dt><dd><tt>run_all_tests<br>
-&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;runs&nbsp;all&nbsp;tests&nbsp;sequentially,&nbsp;if&nbsp;there&nbsp;are&nbsp;any&nbsp;failures&nbsp;the&nbsp;test&nbsp;function&nbsp;docstring&nbsp;is&nbsp;printed</tt></dd></dl>
- <dl><dt><a name="-stats_addanovap_mock1"><strong>stats_addanovap_mock1</strong></a>()</dt><dd><tt>stats_addanovap_mock1<br>
-&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Uses&nbsp;the&nbsp;normalized&nbsp;data&nbsp;from&nbsp;mock_data_1.csv&nbsp;to&nbsp;compute&nbsp;ANOVA&nbsp;p-values&nbsp;for&nbsp;groups&nbsp;"A"&nbsp;and&nbsp;"B"<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Test&nbsp;fails&nbsp;if&nbsp;the&nbsp;shape&nbsp;of&nbsp;the&nbsp;resulting&nbsp;"test_anova"&nbsp;statistic&nbsp;array&nbsp;in&nbsp;the&nbsp;Dataset&nbsp;object&nbsp;does&nbsp;not&nbsp;have&nbsp;<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;the&nbsp;proper&nbsp;shape:&nbsp;(5,)<br>
-&nbsp;&nbsp;&nbsp;&nbsp;returns:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(bool)&nbsp;--&nbsp;test&nbsp;pass&nbsp;(True)&nbsp;or&nbsp;fail&nbsp;(False)</tt></dd></dl>
- <dl><dt><a name="-stats_addpca3loadings_mock1"><strong>stats_addpca3loadings_mock1</strong></a>()</dt><dd><tt>stats_addpca3loadings_mock1<br>
-&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<br>
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Test&nbsp;fails&nbsp;if<br>
 &nbsp;&nbsp;&nbsp;&nbsp;returns:<br>
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(bool)&nbsp;--&nbsp;test&nbsp;pass&nbsp;(True)&nbsp;or&nbsp;fail&nbsp;(False)</tt></dd></dl>
 </td></tr></table>
