+++ lipydomics/doc/lipydomics.test.html
@@ -23,6 +23,8 @@ description:<br>
     
 <tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="lipydomics.test.__main__.html">__main__</a><br>
-</td><td width="25%" valign=top><a href="lipydomics.test.tests.html">tests</a><br>
-</td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table>
+</td><td width="25%" valign=top><a href="lipydomics.test.data.html">data</a><br>
+</td><td width="25%" valign=top><a href="lipydomics.test.plotting.html">plotting</a><br>
+</td><td width="25%" valign=top><a href="lipydomics.test.stats.html">stats</a><br>
+</td></tr></table></td></tr></table>
 </body></html>
