+++ lipydomics/doc/lipydomics.test.plotting.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<html><head><title>Python: module lipydomics.test.plotting</title>
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+</head><body bgcolor="#f0f0f8">
+
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
+<tr bgcolor="#7799ee">
+<td valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong><a href="lipydomics.html"><font color="#ffffff">lipydomics</font></a>.<a href="lipydomics.test.html"><font color="#ffffff">test</font></a>.plotting</strong></big></big></font></td
+><td align=right valign=bottom
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/plotting.py">/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/plotting.py</a></font></td></tr></table>
+    <p><tt>lipydomics/test/plotting.py<br>
+Dylan&nbsp;H.&nbsp;Ross<br>
+2019/09/25<br>
+&nbsp;<br>
+description:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;Define&nbsp;tests&nbsp;associated&nbsp;with&nbsp;the&nbsp;lipydomics/plotting.py&nbsp;module</tt></p>
+
+</body></html>
