+++ lipydomics/doc/lipydomics.test.stats.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<html><head><title>Python: module lipydomics.test.stats</title>
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+</head><body bgcolor="#f0f0f8">
+
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
+<tr bgcolor="#7799ee">
+<td valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong><a href="lipydomics.html"><font color="#ffffff">lipydomics</font></a>.<a href="lipydomics.test.html"><font color="#ffffff">test</font></a>.stats</strong></big></big></font></td
+><td align=right valign=bottom
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/stats.py">/Users/DylanRoss/Documents/GitHub/lipydomics/lipydomics/test/stats.py</a></font></td></tr></table>
+    <p><tt>lipydomics/test/stats.py<br>
+Dylan&nbsp;H.&nbsp;Ross<br>
+2019/09/25<br>
+&nbsp;<br>
+description:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;Define&nbsp;tests&nbsp;associated&nbsp;with&nbsp;the&nbsp;lipydomics/stats.py&nbsp;module</tt></p>
+<p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#aa55cc">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="numpy.html">numpy</a><br>
+</td><td width="25%" valign=top><a href="os.html">os</a><br>
+</td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#eeaa77">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Functions</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#eeaa77"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><dl><dt><a name="-addanovap_mock1"><strong>addanovap_mock1</strong></a>()</dt><dd><tt>addanovap_mock1<br>
+&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Uses&nbsp;the&nbsp;normalized&nbsp;data&nbsp;from&nbsp;mock_data_1.csv&nbsp;to&nbsp;compute&nbsp;ANOVA&nbsp;p-values&nbsp;for&nbsp;groups&nbsp;"A"&nbsp;and&nbsp;"B"<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Test&nbsp;fails&nbsp;if&nbsp;the&nbsp;shape&nbsp;of&nbsp;the&nbsp;resulting&nbsp;"test_anova"&nbsp;statistic&nbsp;array&nbsp;in&nbsp;the&nbsp;Dataset&nbsp;object&nbsp;does&nbsp;not&nbsp;have&nbsp;<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;the&nbsp;proper&nbsp;shape:&nbsp;(5,)<br>
+&nbsp;&nbsp;&nbsp;&nbsp;returns:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(bool)&nbsp;--&nbsp;test&nbsp;pass&nbsp;(True)&nbsp;or&nbsp;fail&nbsp;(False)</tt></dd></dl>
+ <dl><dt><a name="-addpca3_mock1"><strong>addpca3_mock1</strong></a>()</dt><dd><tt>addpca3_mock1<br>
+&nbsp;&nbsp;&nbsp;&nbsp;description:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Uses&nbsp;the&nbsp;raw&nbsp;data&nbsp;from&nbsp;mock_data_1.csv&nbsp;to&nbsp;compute&nbsp;a&nbsp;3&nbsp;component&nbsp;PCA.<br>
+&nbsp;<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Test&nbsp;fails&nbsp;if&nbsp;there&nbsp;are&nbsp;any&nbsp;errors&nbsp;or&nbsp;if&nbsp;the&nbsp;shapes&nbsp;of&nbsp;the&nbsp;following&nbsp;stats&nbsp;entries&nbsp;are&nbsp;incorrect:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;dset.stats['PCA3_loadings_raw'].shape&nbsp;=&nbsp;(3,&nbsp;5)<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;dset.stats['PCA3_projections_raw'].shape&nbsp;=&nbsp;(6,&nbsp;3)<br>
+&nbsp;&nbsp;&nbsp;&nbsp;returns:<br>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(bool)&nbsp;--&nbsp;test&nbsp;pass&nbsp;(True)&nbsp;or&nbsp;fail&nbsp;(False)</tt></dd></dl>
+</td></tr></table>
+</body></html>
