+++ lipydomics/test/__main__.py
@@ -4,11 +4,53 @@
     2019/02/02
 
     description:
-        Calls the run_all_tests method to run all of the tests defined in tests.py
+        Calls the run_all_tests method to run all of the tests defined in the various testing modules
 """
 
 
-from lipydomics.test.tests import run_all_tests
+import traceback
+import os
+
+
+
+
+from lipydomics.test.data import dataset_init_mock1, dataset_normalize_mock1, dataset_getgroup_mock1
+from lipydomics.test.stats import addanovap_mock1, addpca3_mock1
+
+
+def run_all_tests():
+    """
+run_all_tests
+    description:
+        runs all tests sequentially, if there are any failures the test function docstring is printed. A traceback is 
+        printed as well if an exception is thrown.
+"""
+    # references to al of the test functions to be run
+    all_tests = [
+        dataset_init_mock1,
+        dataset_normalize_mock1,
+        dataset_getgroup_mock1,
+        addanovap_mock1,
+        addpca3_mock1
+    ]
+    # run the tests
+    failed = False
+    print("running all tests ... ", end="")
+    for test in all_tests:
+        try:
+            passed = test()
+        except Exception as e:
+            print('\n', test.__doc__, 'TEST FAILED WITH EXEPTION!\n', e)
+            print(traceback.format_exc())
+            failed = True
+            break
+        if not passed:
+            print('\n', test.__doc__, 'TEST FAILED!\n')
+            failed = True
+            break
+    if not failed:
+        print("passed")
+
 
 
 if __name__ == '__main__':
