+++ lipydomics/test/plotting.py
@@ -1,0 +1,10 @@
+"""
+    lipydomics/test/plotting.py
+    Dylan H. Ross
+    2019/09/25
+
+    description:
+        Define tests associated with the lipydomics/plotting.py module
+"""
+
+
